requirements.txt
flask
torch
faiss-cpu
transformers
pillow
matplotlib
